LayoutTests/platform/wk2/http/tests/xmlhttprequest/sync-delegate-callbacks-expected.txt

ALERT: sending synchronous xhr
canAuthenticateAgainstProtectionSpace
ALERT: hello
ALERT: sending synchronous xhr
ALERT: hello
